bedrock-news-japan/requirements.txt
streamlit>=1.15
python-dotenv>=1.0
feedparser>=6.0
boto3>=1.26
